refactor(gateway): share attachment failure logging
